requirements.txt
streamlit
pandas
streamlit_authenticator
streamlit_option_menu
google-cloud-firestore